Made util out of scroll helper in examples
