+++ packages/react-core/src/helpers/index.ts
@@ -13,4 +13,5 @@ export * from './SSRSafeIds/SSRSafeIds';
 export * from './KeyboardHandler';
 export * from './resizeObserver';
 export * from './useInterval';
+export * from './useIsStuckFromScrollParent';
 export * from './datetimeUtils';
